Move fees-related data structures to a separate module
